backend/requirements.txt
# 基础依赖
Flask==2.3.3
Flask-Cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
Flask-Limiter>=3.3.0
pandas>=1.3.0

